simpliy symbol internal data
